django_backend/requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.4
django-database-url==1.0.3
gunicorn==20.1.0
psycopg2==2.9.5
sqlparse==0.4.3
djangorestframework==3.14.0
django-cors-headers==4.0.0
numpy==1.24.3
scikit-surprise==1.1.3
pandas==2.0.1